added darkmode

+++ src/app/profile/profile.component.ts
@@ -6,6 +6,7 @@ import { User } from 'src/models/user.class';
 import { AuthService } from '../services/auth.service';
 import { UserData } from '../../models/user-data';
 import { CrudService } from '../services/crud.service';
+import { DarkmodeService } from '../services/darkmode.service';
 
 @Component({
   selector: 'app-profile',
@@ -15,20 +16,19 @@ import { CrudService } from '../services/crud.service';
 export class ProfileComponent implements OnInit {
   userData: UserData;
   user = new User();
-  constructor(public router: Router,public authService: AuthService, public crud: CrudService, public firestore: Firestore) {
+  constructor(public router: Router, public authService: AuthService, public crud: CrudService, public firestore: Firestore, public dm:DarkmodeService) {
 
   }
-ngOnInit() {
-  this.authService.afAuth.authState.subscribe((user) => {
-    if (user) {
-      this.getUserById()
-      .subscribe((result) => {
-        console.log(result);
-        this.userData = result; 
-        console.log(this.userData['email'])
-      })
-    } 
-  });
+  ngOnInit() {
+    this.authService.afAuth.authState.subscribe((user) => {
+      if (user) {
+        this.getUserById()
+          .subscribe((result) => {
+            console.log(result);
+            this.userData = result;
+          })
+      }
+    });
 
   }
 
@@ -38,8 +38,15 @@ ngOnInit() {
   }
 
   deleteUser() {
-    const userIdRef = doc(this.firestore, `users/${this.authService.userData.uid}`);
-    deleteDoc(userIdRef);
-    this.router.navigateByUrl('/login');
+    if (this.authService.userData.uid === 'm9vxpi4hVuercy8uvrdKCx1ytQ62') {
+      alert('no permissions to delete the Guest user')
+    }
+    else{
+      const userIdRef = doc(this.firestore, `users/${this.authService.userData.uid}`);
+      deleteDoc(userIdRef);
+      this.router.navigateByUrl('/login');
+    }
+    
+
   }
 }
